+++ test/support/factory.ex
@@ -27,7 +27,7 @@ defmodule Radiator.Factory do
   end
 
   def enclosure_factory do
-    %Radiator.Media.Audio{
+    %Radiator.Media.AudioFile{
       file: %{file_name: "example.mp3", updated_at: ~N[2019-04-24 09:41:47]},
       byte_length: 123,
       mime_type: "audio/mpeg"
